+++ utils/repository/repository.go
@@ -38,9 +38,9 @@ type Repository interface {
 	GetObrabyID(ctx context.Context, obraID int64) (*entity.Obra, error)
 	SaveEtapa(ctx context.Context, nombre string) error
 	GetEtapabyName(ctx context.Context, nombre string) (*entity.Etapa, error)
-	SavePiso(ctx context.Context, number int64) error
+	SavePiso(ctx context.Context, number int) error
 	GetPisos(ctx context.Context) ([]entity.Piso, error)
-	GetPisobyNumber(ctx context.Context, number int64) (*entity.Piso, error)
+	GetPisobyNumber(ctx context.Context, number int) (*entity.Piso, error)
 	GetPisobyID(ctx context.Context, ID int64) (*entity.Piso, error)
 	GetObraPisos(ctx context.Context, obraID int64) ([]entity.PisoObra, error)
 	SaveObraPiso(ctx context.Context, obraID, pisoID int64) error
